+++ semver/cmp.ts
@@ -1,6 +1,5 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { SemVer } from "./semver.ts";
-import { Operator } from "./types.ts";
+import type { Operator, SemVer } from "./types.ts";
 import { eq } from "./eq.ts";
 import { neq } from "./neq.ts";
 import { gte } from "./gte.ts";
