+++ semver/difference.ts
@@ -1,6 +1,5 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { SemVer } from "./semver.ts";
-import { ReleaseType } from "./types.ts";
+import type { ReleaseType, SemVer } from "./types.ts";
 import { parse } from "./parse.ts";
 import { eq } from "./eq.ts";
