+++ semver/eq.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 import { compare } from "./compare.ts";
-import { SemVer } from "./semver.ts";
+import type { SemVer } from "./types.ts";
 
 /** Returns `true` if they're logically equivalent, even if they're not the exact
  * same version object. */
