+++ semver/is_valid_operator.ts
@@ -1,0 +1,27 @@
+import type { Operator } from "./types.ts";
+
+/**
+ * Checks to see if the value is a valid Operator string.
+ *
+ * Adds a type assertion if true.
+ * @param value The value to check
+ * @returns True if the value is a valid Operator string otherwise false.
+ */
+export function isValidOperator(value: unknown): value is Operator {
+  if (typeof value !== "string") return false;
+  switch (value) {
+    case "":
+    case "=":
+    case "==":
+    case "===":
+    case "!==":
+    case "!=":
+    case ">":
+    case ">=":
+    case "<":
+    case "<=":
+      return true;
+    default:
+      return false;
+  }
+}
