+++ semver/is_valid_string.ts
@@ -1,0 +1,18 @@
+export const MAX_LENGTH = 256;
+
+/**
+ * Returns true if the value is a valid semver pre-release or build identifier.
+ *
+ * Must be a string. Must be between 1 and 256 characters long. Must match
+ * the regular expression /[0-9A-Za-z-]+/.
+ * @param value The value to check
+ * @returns True if the value is a valid semver string.
+ */
+export function isValidString(value: unknown): value is string {
+  return (
+    typeof value === "string" &&
+    value.length > 0 &&
+    value.length <= MAX_LENGTH &&
+    !!value.match(/[0-9A-Za-z-]+/)
+  );
+}
