+++ semver/range_max.ts
@@ -1,0 +1,21 @@
+// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+import { INVALID } from "./semver.ts";
+import { sort } from "./sort.ts";
+import type { SemVer, SemVerRange } from "./types.ts";
+import { rangeTest } from "./range.ts";
+
+/**
+ * The maximum valid SemVer for a given range or INVALID
+ * @param range The range to calculate the max for
+ * @returns A valid SemVer or INVALID
+ */
+export function rangeMax(range: SemVerRange): SemVer | undefined {
+  // For and's, you take the smallest max
+  // For or's, you take the biggest max
+  //[ [1 and 2] or [2 and 3] ] = [ 1 or 2 ] = 2
+  return sort(
+    range.ranges.map((r) =>
+      sort(r.filter((c) => rangeTest(c.max, range)).map((c) => c.max)).shift()!
+    ),
+  ).filter((v) => v).pop() ?? INVALID;
+}
